Put all characteristics on every animal and changed types of CharacteristicsOfAnimal.weight to double, and CharacteristicsOfAnimal.amountOfKgForFullSaturation to double

+++ src/java/com/javarush/island/Annotations/CharacteristicsOfAnimal.java
@@ -8,8 +8,8 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CharacteristicsOfAnimal {
-    long weight();
-    long amountOfKgForFullSaturation();
+    double weight();
+    double amountOfKgForFullSaturation();
     int walkingDistance();
     int maxQuantityOfAnimal();
 }
